refactor(settings): extract GeneralSection and collapse SettingsModal to shell
Moves the largest tab — storage (notes directory), plain backup &
restore, encrypted AES-256 backup + import flow, auto-lock timeout,
auto-save delay + indicator, and clear-all-notes danger zone — into
its own file under src/components/settings/sections/GeneralSection.tsx.

SettingsModal.tsx is now a 212-LoC tab shell: header, tab bar, and
routing to per-tab section components. All nine tabs still render
(General, Appearance, Editor, Features, Sidebar, Calendar, Templates,
Data, About). The Data tab continues to use SettingsData.

Shared UI primitives live under ./common/ (Toggle, InfoTooltip,
ShortcutRow). IPC calls go through @/lib wrapper functions, which
internally route through safeInvoke from @/lib/ipc — no raw invoke.

No feature removal. Prop signatures for SettingsModal unchanged.
